Moved some unittests, edited setup.py entry_point modules
