+++ src/openqaoa-core/openqaoa/backends/plugin_finder.py
@@ -17,6 +17,7 @@ def plugin_finder_dict() -> list:
         for each_plugin_entry_point in available_plugins:
             try:
                 output_dict[each_plugin_entry_point.name] = each_plugin_entry_point.load()
+                print(output_dict)
             except ModuleNotFoundError:
                 print("The {} module has not been installed.".format(each_plugin_entry_point.name))
             except AttributeError:
